pyproject.toml
[project]
name = "manin-main"
version = "0.1.0"
description = ""
authors = [
    { name = "Your Name", email = "you@example.com" }
]
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "manim (>=0.19.0,<0.20.0)",
    "moviepy (>=2.2.1,<3.0.0)",
    "gtts (>=2.5.4,<3.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "pydub (>=0.25.1,<0.26.0)",
    "soundfile (>=0.12.1,<0.14.0)",
    "torch (>=2.8.0,<3.0.0)",
    "torchvision (>=0.23.0,<0.24.0)",
    "torchaudio (>=2.8.0,<3.0.0)",
    "coqui-tts (>=0.27.2,<0.28.0)",
    "kokoro (>=0.9.4,<0.10.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "fastapi (>=0.118.2,<0.119.0)",
    "uvicorn (>=0.37.0,<0.38.0)",
    "cohere (>=5.20.0,<6.0.0)",
    "google-auth (<2.42.0)",
    "google-auth-oauthlib (>=1.2.3,<2.0.0)",
    "google-api-python-client (>=2.187.0,<3.0.0)",
    "google-auth-httplib2 (>=0.2.1,<0.3.0)",
    "langchain-cohere (>=0.5.0,<0.6.0)",
    "langchain-core (>=1.0.4,<2.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-cov = "^7.0.0"

[tool.poetry.dependencies]
torch = { source = "pypi" }
torchvision = { source = "pypi" }
torchaudio = { source = "pypi" }

# ✅ Use the default PyPI implicitly; no need to redefine it.
# ✅ Add PyTorch CUDA source for GPU compatibility.
[[tool.poetry.source]]
name = "torch-cu121"
url = "https://download.pytorch.org/whl/cu121"
priority = "supplemental"

[tool.pytest.ini_options]
# Add src/ so "from parsers..." works
pythonpath = ["src"]
addopts = "-v --cov=src --cov-report=term-missing"
testpaths = ["tests"]
